+++ app/assets/javascripts/discourse/app/components/sidebar/edit-navigation-modal-form/modal.hbs
@@ -1,0 +1,58 @@
+{{#in-element this.modalHeaderAfterTitleElement}}
+  <p class="sidebar__edit-navigation-modal-form__deselect-wrapper">
+    <DButton
+      @class="btn-flat sidebar__edit-navigation-modal-form__deselect-button"
+      @label="sidebar.edit_navigation_modal_form.deselect_button_text"
+      @ariaLabel="sidebar.edit_navigation_modal_form.deselect_button_text"
+      @action={{@deselectAll}}
+    />
+
+    {{@deselectAllText}}
+  </p>
+{{/in-element}}
+
+<div class="sidebar__edit-navigation-modal-form">
+  <DModalBody
+    @title={{@title}}
+    @class="sidebar__edit-navigation-modal-form__body"
+  >
+    <div class="sidebar__edit-navigation-modal-form__filter">
+      <div class="sidebar__edit-navigation-modal-form__filter-input">
+        {{d-icon
+          "search"
+          class="sidebar__edit-navigation-modal-form__filter-input-icon"
+        }}
+
+        <Input
+          class="sidebar__edit-navigation-modal-form__filter-input-field"
+          placeholder={{@inputFilterPlaceholder}}
+          @type="text"
+          @value={{this.filter}}
+          {{on "input" (action @onFilterInput value="target.value")}}
+          autofocus="true"
+        />
+      </div>
+    </div>
+
+    {{yield}}
+  </DModalBody>
+</div>
+
+<div class="modal-footer sidebar__edit-navigation-modal-form__footer">
+  <DButton
+    @class="btn-primary sidebar__edit-navigation-modal-form__save-button"
+    @label="save"
+    @disabled={{@saving}}
+    @action={{@save}}
+  />
+
+  {{#if @showResetDefaultsButton}}
+    <DButton
+      @icon="undo"
+      @class="btn-flat btn-text sidebar__edit-navigation-modal-form__reset-defaults-button"
+      @label="sidebar.edit_navigation_modal_form.reset_to_defaults"
+      @disabled={{@saving}}
+      @action={{@resetToDefaults}}
+    />
+  {{/if}}
+</div>
