+++ app/assets/javascripts/discourse/app/components/sidebar/edit-navigation-modal-form/tags-form.hbs
@@ -1,0 +1,52 @@
+<Sidebar::EditNavigationModalForm::Modal
+  @title="sidebar.tags_form_modal.title"
+  @saving={{this.saving}}
+  @save={{this.save}}
+  @showResetDefaultsButton={{(gt
+    this.siteSettings.default_navigation_menu_tags.length 0
+  )}}
+  @resetToDefaults={{this.resetToDefaults}}
+  @deselectAll={{this.deselectAll}}
+  @deselectAllText={{i18n "sidebar.tags_form_modal.subtitle.text"}}
+  @inputFilterPlaceholder={{i18n "sidebar.tags_form_modal.filter_placeholder"}}
+  @onFilterInput={{this.onFilterInput}}
+>
+  <form class="sidebar-tags-form">
+    {{#if this.tagsLoading}}
+      <div class="spinner"></div>
+    {{else}}
+      {{#if (gt this.filteredTags.length 0)}}
+        {{#each this.filteredTags as |tag|}}
+          <div class="sidebar-tags-form__tag" data-tag-name={{tag.name}}>
+            <Input
+              id={{concat "sidebar-tags-form__input--" tag.name}}
+              class="sidebar-tags-form__input"
+              @type="checkbox"
+              @checked={{includes this.selectedTags tag.name}}
+              {{on "click" (action "toggleTag" tag.name)}}
+            />
+
+            <label
+              class="sidebar-tags-form__tag-label"
+              for={{concat "sidebar-tags-form__input--" tag.name}}
+            >
+              <p>
+                <span class="sidebar-tags-form__tag-label-name">
+                  {{tag.name}}
+                </span>
+
+                <span class="sidebar-tags-form__tag-label-count">
+                  ({{tag.count}})
+                </span>
+              </p>
+            </label>
+          </div>
+        {{/each}}
+      {{else}}
+        <div class="sidebar-tags-form__no-tags">
+          {{i18n "sidebar.tags_form_modal.no_tags"}}
+        </div>
+      {{/if}}
+    {{/if}}
+  </form>
+</Sidebar::EditNavigationModalForm::Modal>
