+++ app/assets/javascripts/discourse/app/components/sidebar/edit-navigation-modal-form/tags-form.js
@@ -53,10 +53,6 @@ export default class extends Component {
     }
   }
 
-  get modalHeaderAfterTitleElement() {
-    return document.getElementById("modal-header-after-title");
-  }
-
   @action
   onFilterInput(filter) {
     discourseDebounce(this, this.#performFiltering, filter, INPUT_DELAY);
