--- app/assets/javascripts/discourse/app/components/sidebar/tags-form-modal.hbs
@@ -1,89 +1,0 @@
-{{#in-element this.modalHeaderAfterTitleElement}}
-  <p class="sidebar-tags-form__deselect">
-    <DButton
-      @class="btn-flat sidebar-tags-form__deselect-all-btn"
-      @label="sidebar.tags_form_modal.subtitle.button_text"
-      @ariaLabel="sidebar.tags_form_modal.subtitle.button_text"
-      @action={{this.deselectAll}}
-    />
-
-    {{i18n "sidebar.tags_form_modal.subtitle.text"}}
-  </p>
-{{/in-element}}
-
-<DModalBody
-  @title="sidebar.tags_form_modal.title"
-  @class="sidebar-tags-form-modal"
->
-  <form class="sidebar-tags-form">
-    <div class="sidebar-tags-form__filter">
-      {{d-icon "search" class="sidebar-tags-form__filter-input-icon"}}
-
-      <Input
-        class="sidebar-tags-form__filter-input-field"
-        placeholder={{i18n "sidebar.tags_form_modal.filter_placeholder"}}
-        @type="text"
-        @value={{this.filter}}
-        {{on "input" (action "onFilterInput" value="target.value")}}
-        autofocus="true"
-      />
-    </div>
-
-    {{#if this.tagsLoading}}
-      <div class="spinner"></div>
-    {{else}}
-      {{#if (gt this.filteredTags.length 0)}}
-        {{#each this.filteredTags as |tag|}}
-          <div class="sidebar-tags-form__tag" data-tag-name={{tag.name}}>
-            <Input
-              id={{concat "sidebar-tags-form__input--" tag.name}}
-              class="sidebar-tags-form__input"
-              @type="checkbox"
-              @checked={{includes this.selectedTags tag.name}}
-              {{on "click" (action "toggleTag" tag.name)}}
-            />
-
-            <label
-              class="sidebar-tags-form__tag-label"
-              for={{concat "sidebar-tags-form__input--" tag.name}}
-            >
-              <p>
-                <span class="sidebar-tags-form__tag-label-name">
-                  {{tag.name}}
-                </span>
-
-                <span class="sidebar-tags-form__tag-label-count">
-                  ({{tag.count}})
-                </span>
-              </p>
-            </label>
-          </div>
-        {{/each}}
-      {{else}}
-        <div class="sidebar-tags-form__no-tags">
-          {{i18n "sidebar.tags_form_modal.no_tags"}}
-        </div>
-      {{/if}}
-    {{/if}}
-  </form>
-</DModalBody>
-
-<div class="modal-footer sidebar-tags-form__modal-footer">
-  <DButton
-    @class="btn-primary sidebar-tags-form__save-button"
-    @label="save"
-    @disabled={{this.saving}}
-    @action={{this.save}}
-  />
-
-  {{#if (gt this.siteSettings.default_navigation_menu_tags.length 0)}}
-    <DButton
-      @icon="undo"
-      @class="btn-flat btn-text sidebar-tags-form__reset-to-defaults-btn"
-      @label="sidebar.tags_form_modal.reset_to_defaults"
-      @disabled={{this.saving}}
-      @action={{this.resetToDefaults}}
-    />
-  {{/if}}
-
-</div>
