+++ app/assets/javascripts/discourse/app/templates/modal/sidebar-categories-form.hbs
@@ -1,1 +1,3 @@
-<Sidebar::CategoriesFormModal @closeModal={{(action "closeModal")}} />
+<Sidebar::EditNavigationModalForm::CategoriesForm
+  @closeModal={{(action "closeModal")}}
+/>
