+++ app/assets/javascripts/discourse/app/templates/modal/sidebar-tags-form.hbs
@@ -1,1 +1,3 @@
-<Sidebar::TagsFormModal @closeModal={{(action "closeModal")}} />
+<Sidebar::EditNavigationModalForm::TagsForm
+  @closeModal={{(action "closeModal")}}
+/>
