--- app/assets/stylesheets/common/components/sidebar-tags-form.scss
@@ -1,93 +1,0 @@
-.sidebar-tags-form-modal {
-  .modal-body {
-    min-height: 30vh;
-  }
-}
-
-.sidebar-tags-form {
-  display: flex;
-  flex-wrap: wrap;
-
-  .sidebar-tags-form__tag {
-    flex-basis: 30%;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    padding: 0.5em 0.25em;
-
-    .sidebar-tags-form__tag-label {
-      margin-bottom: 0;
-
-      p {
-        margin: 0;
-      }
-    }
-
-    .sidebar-tags-form__input {
-      margin-top: 0;
-    }
-  }
-
-  .sidebar-tags-form__tag-label-count {
-    color: var(--primary-medium);
-  }
-
-  .sidebar-tags-form__filter {
-    display: flex;
-    flex-direction: row;
-    margin-right: auto;
-    width: 100%;
-    position: relative;
-  }
-
-  .sidebar-tags-form__filter-input-icon {
-    position: absolute;
-    left: 0.5em;
-    top: 0.65em;
-    color: var(--primary-low-mid);
-  }
-
-  .sidebar-tags-form__filter-input-field {
-    border-color: var(--primary-low-mid);
-    padding-left: 1.75em;
-    width: 100%;
-
-    &:focus {
-      border-color: var(--tertiary);
-      outline: none;
-      outline-offset: 0;
-      box-shadow: inset 0px 0px 0px 1px var(--tertiary);
-    }
-  }
-}
-
-.sidebar-tags-form__modal-footer.modal-footer {
-  .sidebar-tags-form__reset-to-defaults-btn {
-    margin-left: auto;
-    margin-right: 0em;
-
-    .d-icon {
-      font-size: var(--font-down-1);
-      color: var(--tertiary);
-    }
-
-    &:focus,
-    &:active {
-      background: none;
-    }
-  }
-}
-
-.sidebar-tags-form__deselect {
-  margin: 0;
-
-  .btn-flat.sidebar-tags-form__deselect-all-btn {
-    margin-left: auto;
-    padding: 0;
-
-    &:focus,
-    &:active {
-      background: none;
-    }
-  }
-}
