+++ app/assets/stylesheets/common/components/sidebar/edit-navigation-modal-form/categories-form.scss
@@ -1,69 +1,4 @@
-.sidebar-categories-form-modal {
-  .modal-body {
-    min-height: 50vh;
-  }
-}
-
-.sidebar-categories-form__modal-footer.modal-footer {
-  .sidebar-categories-form__reset-to-defaults-btn {
-    margin-left: auto;
-    margin-right: 0em;
-
-    .d-icon {
-      font-size: var(--font-down-1);
-      color: var(--tertiary);
-    }
-
-    &:focus,
-    &:active {
-      background: none;
-    }
-  }
-}
-
-.sidebar-categories-form__deselect {
-  margin: 0;
-
-  .btn-flat.sidebar-categories-form__deselect-all-btn {
-    margin-left: auto;
-    padding: 0;
-
-    &:focus,
-    &:active {
-      background: none;
-    }
-  }
-}
-
 .sidebar-categories-form {
-  .sidebar-categories-form__filter {
-    display: flex;
-    flex-direction: row;
-    margin-right: auto;
-    width: 100%;
-    position: relative;
-  }
-
-  .sidebar-categories-form__filter-input-icon {
-    position: absolute;
-    left: 0.5em;
-    top: 0.65em;
-    color: var(--primary-low-mid);
-  }
-
-  .sidebar-categories-form__filter-input-field {
-    border-color: var(--primary-low-mid);
-    padding-left: 1.75em;
-    width: 100%;
-
-    &:focus {
-      border-color: var(--tertiary);
-      outline: none;
-      outline-offset: 0;
-      box-shadow: inset 0px 0px 0px 1px var(--tertiary);
-    }
-  }
-
   .sidebar-categories-form__row {
     display: flex;
     flex-direction: column;
