+++ app/assets/stylesheets/common/components/sidebar/edit-navigation-modal-form/modal.scss
@@ -1,0 +1,71 @@
+.sidebar__edit-navigation-modal-form {
+  .modal-body {
+    min-height: 50vh;
+  }
+
+  .sidebar__edit-navigation-modal-form__filter {
+    display: flex;
+    flex-direction: row;
+    margin-bottom: 1em;
+    width: 100%;
+
+    .sidebar__edit-navigation-modal-form__filter-input {
+      display: flex;
+      flex-direction: row;
+      margin-right: auto;
+      width: 100%;
+      position: relative;
+    }
+
+    .sidebar__edit-navigation-modal-form__filter-input-icon {
+      position: absolute;
+      left: 0.5em;
+      top: 0.65em;
+      color: var(--primary-low-mid);
+    }
+
+    .sidebar__edit-navigation-modal-form__filter-input-field {
+      border-color: var(--primary-low-mid);
+      padding-left: 1.75em;
+      width: 100%;
+
+      &:focus {
+        border-color: var(--tertiary);
+        outline: none;
+        outline-offset: 0;
+        box-shadow: inset 0px 0px 0px 1px var(--tertiary);
+      }
+    }
+  }
+}
+
+.sidebar__edit-navigation-modal-form__footer {
+  .btn.sidebar__edit-navigation-modal-form__reset-defaults-button {
+    margin-left: auto;
+    margin-right: 0em;
+
+    .d-icon {
+      font-size: var(--font-down-1);
+      color: var(--tertiary);
+    }
+
+    &:focus,
+    &:active {
+      background: none;
+    }
+  }
+}
+
+.sidebar__edit-navigation-modal-form__deselect-wrapper {
+  margin: 0;
+
+  .btn-flat.sidebar__edit-navigation-modal-form__deselect-button {
+    margin-left: auto;
+    padding: 0;
+
+    &:focus,
+    &:active {
+      background: none;
+    }
+  }
+}
