+++ app/assets/stylesheets/desktop/components/_index.scss
@@ -1,5 +1,5 @@
-@import "sidebar-categories-form";
-@import "sidebar-tags-form";
+@import "sidebar/edit-navigation-modal-form/categories-form";
+@import "sidebar/edit-navigation-modal-form/tags-form";
 @import "user-card";
 @import "user-info";
 @import "user-stream-item";
