+++ app/assets/stylesheets/desktop/components/sidebar/edit-navigation-modal-form/categories-form.scss
