+++ app/assets/stylesheets/desktop/components/sidebar/edit-navigation-modal-form/tags-form.scss
