+++ app/assets/stylesheets/mobile/components/_index.scss
@@ -1,5 +1,5 @@
-@import "sidebar-categories-form";
-@import "sidebar-tags-form";
+@import "sidebar/edit-navigation-modal-form/modal";
+@import "sidebar/edit-navigation-modal-form/tags-form";
 @import "topic-footer-mobile-dropdown";
 @import "user-card";
 @import "user-stream-item";
