+++ app/assets/stylesheets/mobile/components/sidebar/edit-navigation-modal-form/modal.scss
@@ -1,4 +1,4 @@
-.sidebar-categories-form-modal {
+.sidebar__edit-navigation-modal-form {
   .modal-inner-container {
     width: 35em;
   }
