+++ app/assets/stylesheets/mobile/components/sidebar/edit-navigation-modal-form/tags-form.scss
